Refactor: 검색기능 새로고침이 아닌 상태변화로 수정
